Move client test loop to the test suite
