+++ lspTest/src/test/scala/lspTest/TestSuite.scala
@@ -1,14 +1,46 @@
 package lspTest
 
+import typings.node.fsMod
+
 import scala.concurrent.{ExecutionContext, Future}
-import utest._
+import utest.*
 
 object TestSuite extends TestSuite {
+  def samplesDir = "examples/benchmarks"
+
   // Provide a connection to the Effekt language server to the test case
   def withServer[T](testBody: ServerConnection => Future[T])
                    (implicit ec: ExecutionContext): Future[T] = {
     val serverConn = new ServerConnection
     testBody(serverConn).andThen { case _ => serverConn.cleanup() }
+  }
+
+  def forEachSample(callback: String => Future[Unit])(implicit ec: ExecutionContext): Future[Unit] = {
+    forEachFile(samplesDir, callback)
+  }
+
+  def forEachFile(path: String, callback: String => Future[Unit])(implicit ec: ExecutionContext): Future[Unit] = {
+    callback("examples/benchmarks/are_we_fast_yet/bounce.effekt")
+    //    linearizeFuture {
+    //      fsMod.readdirSync(path).toSeq.flatMap { sub =>
+    //        val subPath = s"$path/$sub"
+    //        if (fsMod.statSync(subPath).get.isDirectory()) {
+    //          Seq(() => forEachFile(subPath, callback)) // iterate in sub directory
+    //        } else if (subPath.endsWith(".effekt")) {
+    //          Seq(() => callback(subPath))
+    //        } else {
+    //          Seq.empty[() => Future[Unit]]
+    //        }
+    //      }
+    //    }.map(_ => ())
+  }
+
+  // compared to Future.sequence, this function will also *start* the futures sequentially and not concurrently
+  def linearizeFuture(in: Seq[() => Future[Unit]])(implicit ec: ExecutionContext): Future[Seq[Unit]] = {
+    val builder = Seq.newBuilder[Unit]
+    in.foldLeft(Future.successful(())) {
+      (a, b) => a.transformWith { _ => b() }
+    } map (_ => builder.result())
   }
 
   def tests = Tests {
@@ -41,6 +73,22 @@ object TestSuite extends TestSuite {
           _ <- client.shutdown()
           _ <- client.exit()
         } yield assert(serverConn.hasExited)
+      }
+    }
+
+    test("Run all client tests") {
+      forEachSample { file =>
+        println(s"Running language server tests for $file")
+        withServer { serverConn =>
+          for {
+            client <- serverConn.setupServerAndConnect()
+            tests = new TestSingleFile(client, file)
+            results <- TestRunner.runAndPrintAsync(tests.tests, "Tests", executor = SequentialExecutor)
+          } yield {
+            // use Predef to not repeat all previous error messages
+            Predef.assert(results.leaves.forall(leaf => leaf.value.isSuccess), "some client tests failed")
+          }
+        }
       }
     }
   }
